Add files via upload
An app that can store the notes of user. It has interactive interface and due to its responsive design, desktop user as well as mobile user can use this app. It has a random quote generator and a slideshow so, user can feel motivated and fresh while working on app.
•	This app is created using HTML, CSS and JavaScript.
•	During this project I’ve learned how we can interact with DOM elements and manipulate according to user input. Also learned to store user input in local Storage. 
•	I’ve inbuilt a search bar also, using the conditional function so, It’ll be easier to find the created note.